Moved i18n-folder to presenter layer
