+++ src/main/js/presenter/i18n/i18nConfig.js
